backend/rollups/cartesi/backend/requirements.txt
requests == 2.23.0
numpy == 1.24.2
pycryptodome == 3.*